+++ src/modules/customers/search/modules/SearchCustomer.jsx
@@ -1,7 +1,7 @@
 import { Container, Typography } from "@mui/material";
 import React from "react";
 
-export default function BuscarCliente() {
+export default function SearchCustomer() {
   return(
     <Container maxWidth="lg" style={{display: 'flex', alignItems: 'center', flexDirection: 'column'}}>
       <Typography variant="h3" style={{margin: '10px'}}> Buscar cliente </Typography>
