+++ src/modules/dashboard/modules/Dashboard.jsx
@@ -1,7 +1,7 @@
 import { Container, Typography } from "@mui/material";
 import React from "react";
 
-export default function Tablero() {
+export default function Dashboard() {
   return(
     <Container maxWidth="lg" style={{display: 'flex', alignItems: 'center', flexDirection: 'column'}}>
       <Typography variant="h3" style={{margin: '10px'}}> Tablero </Typography>
